requirements.txt
python-dotenv
Flask
requests
openai